move ica related logic to neutron mod
