+++ e2e/examples/eth_vault/neutron.rs
@@ -555,3 +555,183 @@ pub fn log_neutron_acc_balances(
     info!("POSITION ACC BAL\t: {:?}", position_acc_bal);
     info!("WITHDRAW ACC BAL\t: {:?}", withdraw_acc_bal);
 }
+
+pub mod ica {
+    use std::{error::Error, time::Duration};
+
+    use cosmwasm_std::{Uint128, Uint64};
+    use localic_std::modules::cosmwasm::{contract_execute, contract_instantiate, contract_query};
+    use localic_utils::{
+        utils::test_context::TestContext, DEFAULT_KEY, NEUTRON_CHAIN_ADMIN_ADDR,
+        NEUTRON_CHAIN_DENOM, NEUTRON_CHAIN_NAME,
+    };
+    use log::info;
+    use valence_account_utils::ica::{IcaState, RemoteDomainInfo};
+    use valence_e2e::utils::{
+        ibc::poll_for_ica_state,
+        manager::{ICA_IBC_TRANSFER_NAME, INTERCHAIN_ACCOUNT_NAME},
+        GAS_FLAGS, NOBLE_CHAIN_NAME, UUSDC_DENOM,
+    };
+    use valence_ica_ibc_transfer::msg::RemoteChainInfo;
+    use valence_library_utils::LibraryAccountType;
+
+    pub fn instantiate_interchain_account_contract(
+        test_ctx: &TestContext,
+    ) -> Result<String, Box<dyn Error>> {
+        let ica_account_code = *test_ctx
+            .get_chain(NEUTRON_CHAIN_NAME)
+            .contract_codes
+            .get(INTERCHAIN_ACCOUNT_NAME)
+            .unwrap();
+
+        info!("Instantiating the ICA contract...");
+        let timeout_seconds = 90;
+        let ica_instantiate_msg = valence_account_utils::ica::InstantiateMsg {
+            admin: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+            approved_libraries: vec![],
+            remote_domain_information: RemoteDomainInfo {
+                connection_id: test_ctx
+                    .get_connections()
+                    .src(NEUTRON_CHAIN_NAME)
+                    .dest(NOBLE_CHAIN_NAME)
+                    .get(),
+                ica_timeout_seconds: Uint64::new(timeout_seconds),
+            },
+        };
+
+        let valence_ica = contract_instantiate(
+            test_ctx
+                .get_request_builder()
+                .get_request_builder(NEUTRON_CHAIN_NAME),
+            DEFAULT_KEY,
+            ica_account_code,
+            &serde_json::to_string(&ica_instantiate_msg)?,
+            "valence_ica",
+            None,
+            "",
+        )?;
+        info!(
+            "ICA contract instantiated. Address: {}",
+            valence_ica.address
+        );
+
+        Ok(valence_ica.address)
+    }
+
+    pub fn register_interchain_account(
+        test_ctx: &mut TestContext,
+        interchain_account_addr: &str,
+    ) -> Result<String, Box<dyn Error>> {
+        info!("Registering the ICA...");
+        contract_execute(
+            test_ctx
+                .get_request_builder()
+                .get_request_builder(NEUTRON_CHAIN_NAME),
+            interchain_account_addr,
+            DEFAULT_KEY,
+            &serde_json::to_string(&valence_account_utils::ica::ExecuteMsg::RegisterIca {})
+                .unwrap(),
+            &format!("{} --amount=100000000{}", GAS_FLAGS, NEUTRON_CHAIN_DENOM),
+        )
+        .unwrap();
+        std::thread::sleep(Duration::from_secs(3));
+
+        // We want to check that it's in state created
+        poll_for_ica_state(test_ctx, interchain_account_addr, |state| {
+            matches!(state, IcaState::Created(_))
+        });
+
+        // Get the remote address
+        let ica_state: IcaState = serde_json::from_value(
+            contract_query(
+                test_ctx
+                    .get_request_builder()
+                    .get_request_builder(NEUTRON_CHAIN_NAME),
+                interchain_account_addr,
+                &serde_json::to_string(&valence_account_utils::ica::QueryMsg::IcaState {}).unwrap(),
+            )["data"]
+                .clone(),
+        )
+        .unwrap();
+
+        let remote_address = match ica_state {
+            IcaState::Created(ica_info) => ica_info.address,
+            _ => {
+                unreachable!("Expected IcaState::Created variant");
+            }
+        };
+        info!("Remote address created: {}", remote_address);
+
+        Ok(remote_address)
+    }
+
+    pub fn setup_ica_ibc_transfer_lib(
+        test_ctx: &mut TestContext,
+        interchain_account_addr: &str,
+        amount_to_transfer: u128,
+    ) -> Result<String, Box<dyn Error>> {
+        let ica_ibc_transfer_lib_code = *test_ctx
+            .get_chain(NEUTRON_CHAIN_NAME)
+            .contract_codes
+            .get(ICA_IBC_TRANSFER_NAME)
+            .unwrap();
+
+        info!("ica ibc transfer lib code: {ica_ibc_transfer_lib_code}");
+
+        info!("Instantiating the ICA IBC transfer contract...");
+        let ica_ibc_transfer_instantiate_msg = valence_library_utils::msg::InstantiateMsg::<
+            valence_ica_ibc_transfer::msg::LibraryConfig,
+        > {
+            owner: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+            processor: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+            config: valence_ica_ibc_transfer::msg::LibraryConfig {
+                input_addr: LibraryAccountType::Addr(interchain_account_addr.to_string()),
+                amount: Uint128::new(amount_to_transfer),
+                denom: UUSDC_DENOM.to_string(),
+                receiver: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+                remote_chain_info: RemoteChainInfo {
+                    channel_id: test_ctx
+                        .get_transfer_channels()
+                        .src(NOBLE_CHAIN_NAME)
+                        .dest(NEUTRON_CHAIN_NAME)
+                        .get(),
+                    ibc_transfer_timeout: None,
+                },
+            },
+        };
+
+        let ica_ibc_transfer = contract_instantiate(
+            test_ctx
+                .get_request_builder()
+                .get_request_builder(NEUTRON_CHAIN_NAME),
+            DEFAULT_KEY,
+            ica_ibc_transfer_lib_code,
+            &serde_json::to_string(&ica_ibc_transfer_instantiate_msg)?,
+            "valence_ica_ibc_transfer",
+            None,
+            "",
+        )?;
+        info!(
+            "ICA IBC transfer contract instantiated. Address: {}",
+            ica_ibc_transfer.address
+        );
+
+        info!("Approving the ICA IBC transfer library...");
+        contract_execute(
+            test_ctx
+                .get_request_builder()
+                .get_request_builder(NEUTRON_CHAIN_NAME),
+            interchain_account_addr,
+            DEFAULT_KEY,
+            &serde_json::to_string(&valence_account_utils::ica::ExecuteMsg::ApproveLibrary {
+                library: ica_ibc_transfer.address.clone(),
+            })
+            .unwrap(),
+            GAS_FLAGS,
+        )
+        .unwrap();
+        std::thread::sleep(Duration::from_secs(3));
+
+        Ok(ica_ibc_transfer.address)
+    }
+}
